requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.11.0
channels==4.0.0
click==8.1.7
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.8.0
